Fix logic error where I confused min and max.
Fix error in the mapping of NUMA nodes to heaps where we made the assumption that increasing NUMA node indices would map to increasing heap numbers. This is not a correct assumption - on one machine I looked at, NUMA node 0 mapped to heaps 0-63, 1 mapped to 128-195, 2 mapped to 64-127, and 3 mapped to 196-255. We would assume that we could obtain the end of heap number range for NUMA node 0 by taking the start of the range for NUMA node 1 and subtract 1, so we would think that NUMA node 0 covers 0-127, which is incorrect.

The most straightforward way to fix this is to introduce parallel array for the start and end of the heap range associated with a NUMA node.